save app before create alarm for hour 10

+++ app/src/main/java/com/mkarshnas6/karenstudio/karenbalance/MainActivity.kt
@@ -312,7 +312,7 @@ class MainActivity : AppCompatActivity() {
             val db = DBHandler.getDatabase(context = this)
             val InsertTarget = TargetEntity(
                 name = name,
-                price = price.toInt(),
+                price = price.toLong(),
                 date = date.toString(),
                 necessary = isNecessary,
                 img = img
